fix: Fix site navigation behavior issues from manage space portlet - EXO-65306 - Meeds-io/MIPs#68 (#171)
Prior to this change, after performing an action from the site navigation opened from the space management portlet, the refresh is done without any site type and site name data but with the default current ones. After this change, the site name and site type will change whenever the site navigation drawer is opened so that every action will take the right data.