fix: Translated labels not loaded - EXO-73828. (#334)
Before this change, when change language to arabic and go to site administration then on platform settings site click on navigation and choose first node and click on edit node then check the translations available for first label, the translation labels aren't loaded. After this change, the translation is loaded and displayed when choosing Arabic langauge.